+++ ajwah_bloc/lib/src/createStore.dart
@@ -10,19 +10,17 @@ Store _store;
 ///Also you can dynamically **add** or **remove** state and effect
 ///using **addState(BaseState stateInstance)** ,**removeStateByStateName(String stateName)**,
 ///**addEffects(BaseEffect effectInstance)**, **addEffect(EffectCallback callback, {@required String key})**, **removeEffectsByKey(String key)**
-createStore(
-    {List<BaseState> states,
-    List<BaseEffect> effects = const [],
-    bool block = false}) {
-  var fx = () {
-    _store = Store(states);
-    Future.delayed(Duration(seconds: 5));
-    effects.forEach((effect) {
-      _store.addEffects(effect);
-    });
-    return _store;
-  };
-  return block ? fx() : Future.microtask(fx);
+Store createStore({
+  List<BaseState> states,
+  List<BaseEffect> effects = const [],
+}) {
+  _store = Store(states);
+
+  effects.forEach((effect) {
+    _store.addEffects(effect);
+  });
+
+  return _store;
 }
 
 ///return **Store** instance.