+++ ajwah_bloc/test/dynamicState_test.dart
@@ -11,7 +11,7 @@ void main() {
   Store store;
 
   setUpAll(() {
-    store = createStore(states: [CounterState()], block: true);
+    store = createStore(states: [CounterState()]);
   });
 
   tearDownAll(() {