+++ ajwah_bloc_test/CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.1.0
 
 A Dart package that makes testing ajwah_bloc easy.
+
+## 0.1.1
+
+improved `ajwahTest(...log:(emitList){})`
+[log] is an optional callback which is invoked before [expect].
+[log] is called with the emited `list of state`.